move privateKey to implementation
